refactor: BukkitStaticGachaPrizeFactoryをfactoriesパッケージに移動
